restructure location of tests
